Implemented program clock
Added Clock class which contains the program clock

Added wait(), and instance method of Clock which allows the user to specify a function, and number of clock cycles to wait before calling the function. It returns what ever the return value of the passed function was. Requires a unique instance of the Clock class per thread which needs to wait.